third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-scrollbar-expected.txt

This tests scroll gesture events on a scroll bar. The scrollable div below should be slightly scrolled down if successful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS movingDiv.scrollTop is 0
PASS movingDiv.scrollTop is 0
PASS movingDiv.scrollTop is >= 20
PASS movingDiv.scrollTop is >= 85
PASS movingDiv.scrollTop is >= 85
PASS movingDiv.scrollLeft is 0
PASS movingDiv.scrollLeft is 0
PASS movingDiv.scrollLeft is >= 20
PASS movingDiv.scrollLeft is >= 85
PASS movingDiv.scrollLeft is >= 85
PASS successfullyParsed is true

TEST COMPLETE
